move tests to !netfx target group
