+++ src/System.Runtime.Extensions/tests/System.Runtime.Extensions.Tests.csproj
@@ -33,6 +33,7 @@
     <Compile Include="TestHelpers.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)'=='netcoreapp'">
+    <Compile Include="System\AppDomainTests.netcoreapp.cs" />
     <Compile Include="System\Random.netcoreapp.cs" />
     <Compile Include="System\IO\Path.IsPathFullyQualified.netcoreapp.cs" />
     <Compile Include="System\IO\PathTestsBase.netcoreapp.cs" />
